+++ docs/CHANGELOG.md
@@ -1,13 +1,38 @@
 # [CHANGELOG](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased](https://github.com/gdsfactory/gdsfactory/compare/v6.98.0...main)
+## [Unreleased](https://github.com/gdsfactory/gdsfactory/compare/v6.101.1...main)
+
+## [6.101.1](https://github.com/gdsfactory/gdsfactory/compare/v6.101.1...v6.101.0)
+
+- fix kfactory dependency [PR](https://github.com/gdsfactory/gdsfactory/pull/1714)
+
+## [6.101.0](https://github.com/gdsfactory/gdsfactory/compare/v6.101.0...v6.100.0)
+
+- fix git diff gds [PR](https://github.com/gdsfactory/gdsfactory/pull/1712)
+
+## [6.100.0](https://github.com/gdsfactory/gdsfactory/compare/v6.100.0...v6.99.0)
+
+- add `get_polygon_bbox` and `get_polygon_enclosure` that return a shapely polygon [PR](https://github.com/gdsfactory/gdsfactory/pull/1709)
+
+## [6.99.0](https://github.com/gdsfactory/gdsfactory/compare/v6.99.0...v6.98.2)
+
+- improve difftest [PR](https://github.com/gdsfactory/gdsfactory/pull/1703)
+- fix devsim [PR](https://github.com/gdsfactory/gdsfactory/pull/1704)
+- update tidy3d to 2.2.2
+
+## [6.98.2](https://github.com/gdsfactory/gdsfactory/compare/v6.98.2...v6.98.1)
+
+- only use section.insets if they are not (0, 0) [PR](https://github.com/gdsfactory/gdsfactory/commit/9bffa6d84f4fe427e74cd6193c3afdc731bb0deb)
+- fix webapp [PR](https://github.com/gdsfactory/gdsfactory/pull/1701)
+
+## [6.98.1](https://github.com/gdsfactory/gdsfactory/compare/v6.98.1...v6.98.0)
 
 - update kfactory [PR](https://github.com/gdsfactory/gdsfactory/pull/1695)
 - fix tidy3d 2D FDTD sims and update tidy3d [PR](https://github.com/gdsfactory/gdsfactory/pull/1697)
 
 ## [6.98.0](https://github.com/gdsfactory/gdsfactory/compare/v6.98.0...v6.97.1)
 
-- add component overlap [PR](https://github.com/gdsfactory/gdsfactory/pull/1689/)
+- add verification functions to check for component overlap [PR](https://github.com/gdsfactory/gdsfactory/pull/1689/)
 
 ## [6.97.1](https://github.com/gdsfactory/gdsfactory/compare/v6.97.0...v6.97.1)